refactor(adapters): reorganize repositories
