+++ source/interface_adapters/__init__.py
@@ -1,0 +1,1 @@
+"""Interface adapters for the use cases and frameworks/drivers."""
